refactor(#38): rename files to PascalCase
