+++ libs/ui/src/assets/NoTeam_Receive.png
