+++ libs/ui/src/lib/components/ParticipationChart.tsx
